Internal: Context menu tests
