+++ tests/playwright/sanity/section.test.ts
@@ -1,6 +1,7 @@
 import { expect } from '@playwright/test';
 import { parallelTest as test } from '../parallelTest';
 import WpAdminPage from '../pages/wp-admin-page';
+import ContextMenu from '../pages/widgets/context-menu';
 import widgets from '../enums/widgets';
 
 test.describe( 'Section tests', () => {
@@ -17,6 +18,7 @@ test.describe( 'Section tests', () => {
 	test( 'Verify that elements are in the correct order after passing into a new section', async ( { page, apiRequests }, testInfo ) => {
 		// Arrange.
 		const wpAdmin = new WpAdminPage( page, testInfo, apiRequests );
+		const contextMenu = new ContextMenu( page, testInfo );
 		const editor = await wpAdmin.openNewPage(),
 			sectionId1 = await editor.addElement( { elType: 'section' }, 'document' ),
 			sectionId2 = await editor.addElement( { elType: 'section' }, 'document' ),
@@ -30,13 +32,13 @@ test.describe( 'Section tests', () => {
 		await editor.addWidget( widgets.heading, section2ColumnId );
 
 		// Copy section 1.
-		await editor.copyElement( sectionId1 );
+		await contextMenu.copyElement( sectionId1 );
 
 		// Open Add Section Inline element.
 		await editor.openAddElementSection( sectionId2 );
 
 		// Paste section 1 onto New section element.
-		await editor.pasteElement( '.elementor-add-section-inline' );
+		await contextMenu.pasteElement( '.elementor-add-section-inline' );
 
 		// Assert.
 		// Verify that the first section has a `data-id` value of `sectionId1`.
